composite file

+++ fr.parisnanterre.global/TD.Restaurant.Patterns/src/Burger/Ingredient.java
@@ -2,9 +2,9 @@ package Burger;
 
 public interface Ingredient {
 
-    int get_grams();
-    int get_calories();
-    int get_prices();
+    double get_grams();
+    double get_calories();
+    double get_prices();
 
     default String get_print(){
         return "- "+this+" ("+get_grams()+"g) -- "+get_prices()+"€\n";
